# BenefitsEnrollment/src/main/resources/openapi/benefits-enrollment-openapi.yaml
openapi: 3.0.3
info:
  title: Benefits Enrollment Service API
  version: v1
  description: REST API for managing employees, plans, and enrollments with switchable repository backends.
servers:
  - url: http://localhost:8080
...
paths:
...
  /plans:
    get:
      tags: [Plans]
      summary: List plans with pagination and filters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/PlanType'
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPlanResponse'
    post:
      tags: [Plans]
      summary: Create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
      responses:
        '201':
          description: Plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
  /plans/{id}:
    get:
      tags: [Plans]
      summary: Get plan by id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Plan found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Plans]
      summary: Update plan
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanRequest'
      responses:
        '200':
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Plans]
      summary: Delete plan
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Plan deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    PlanType:
      type: string
      enum: [MEDICAL, DENTAL, VISION, LIFE, RETIREMENT]
...
    CreatePlanRequest:
      type: object
      required: [id, name, type, cost]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          $ref: '#/components/schemas/PlanType'
        cost:
          type: number
          format: double
    UpdatePlanRequest:
      type: object
      required: [name, type, cost]
      properties:
        name:
          type: string
        type:
          $ref: '#/components/schemas/PlanType'
        cost:
          type: number
          format: double
    PlanResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          $ref: '#/components/schemas/PlanType'
        cost:
          type: number
          format: double
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        errorCode:
          type: string
        message:
          type: string
...
    PaginatedPlanResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PlanResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...